[plane_seg_ros] Use tf to look-up transform of input; publish in same frame as output
This removes the hard-coded dependency on "odom" when the elevation map
is not produced in the odom frame (e.g. the point cloud odom frame) --
the result was an offset between the data published by plane_seg_ros and
the actual elevation map. By removing a subscriber to the state
estimator, a further bug is fixed: Delayed messages would have out of
sync timesteps with the stored last_robot_pose_.