[plane_seg_ros] Add MarkerArray publishing and move all publishers to lazy mode
This commit makes two changes:
1) It adds a new publisher publishing a visualization_msgs/MarkerArray.
The benefit of this is to have a publisher that publishes all regions
separately as a vector for processing in downstream applications.
2) All code to generate the output for publishing and the message
publishing is now performed lazily: It is only carried out when there is
at least one subscriber to the publisher/topic. This reduces the
computation time by not creating data structures that aren't needed.